Main Searching Code, Rainmeter-Based
First Commit of Rainmeter-based league in-game search.
Basic, functional, "working" code at the moment.

Takes in a summoner's name (currently in game).
Converts it to a unique User ID (Riot's id for the user).
Takes that in to find the summoner's current game and the current
season's ranked statistics for all the people in the current game.

Currently has placeholder text for bad calls to the system.
Also does not automatically process the User ID to the second and third
calls to Riot's API.

Main TODOs:
Account for people not in game.
Do something about filler text.
Make the UI pretty.